FE-1 refactor to typescript
