+++ src/main/front/src/components/Navbar.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Navbar = () => {
+    return <h1>Navbar</h1>;
+}
+
+export default Navbar;
