--- src/main/front/src/components/Navigation.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const NavigationWrapper = styled.nav`
-    position: absolute;
-    
-`;
