+++ src/main/front/src/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom'
+import './index.css';
+import App from './app/App';
+
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+
+root.render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+);
